Xtail POC implementation in Python
Script is able to follow several log files simultaneously, printing the parsed events to the screen. The event separator (currently \n character) can be customized in order to parse multiline events. the output will be sorted (expects the ISO-formated time in the beginning of the event to keep cronological order).